Merged develop/1.2.2 into master

+++ src/main/java/org/thingsboard/gateway/service/DefaultTenantManagerService.java
@@ -17,14 +17,12 @@ package org.thingsboard.gateway.service;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 import org.thingsboard.gateway.extensions.ExtensionService;
 import org.thingsboard.gateway.extensions.http.HttpService;
 import org.thingsboard.gateway.service.conf.TbExtensionConfiguration;
 import org.thingsboard.gateway.service.conf.TbGatewayConfiguration;
 import org.thingsboard.gateway.service.conf.TbTenantConfiguration;
 import org.thingsboard.gateway.service.gateway.GatewayService;
-import org.thingsboard.gateway.service.gateway.MqttGatewayService;
 
 import javax.annotation.PostConstruct;
 import javax.annotation.PreDestroy;
@@ -32,22 +30,24 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.function.Consumer;
 
 /**
  * Created by ashvayka on 29.09.17.
  */
-@Service
 @Slf4j
-public class DefaultTenantManagerService implements TenantManagerService {
+public abstract class DefaultTenantManagerService implements TenantManagerService {
 
     @Autowired
     private TbGatewayConfiguration configuration;
 
-    private Map<String, TenantServicesRegistry> gateways;
+    private Map<String, TenantServiceRegistry> gateways;
     private List<HttpService> httpServices;
     private Boolean isRemoteConfiguration;
 
     private static final String STATUS_STOP = "Stopped";
+
+    public abstract GatewayService getGatewayService(TbTenantConfiguration configuration, Consumer<String> extensionsConfigListener);
 
     @PostConstruct
     public void init() {
@@ -58,14 +58,18 @@ public class DefaultTenantManagerService implements TenantManagerService {
             if (isRemoteConfiguration) {
                 String label = configuration.getLabel();
                 log.info("[{}] Initializing gateway", configuration.getLabel());
-                GatewayService service = new MqttGatewayService(configuration, c -> onExtensionConfigurationUpdate(label, c));
+                TenantServiceRegistry tenantServiceRegistry = new TenantServiceRegistry();
+                GatewayService service = null;
                 try {
-                    service.init();
-                    gateways.put(label, new TenantServicesRegistry(service));
+                    service = getGatewayService(configuration, c -> tenantServiceRegistry.updateExtensionConfiguration(c));
+                    tenantServiceRegistry.setService(service);
+                    gateways.put(label, tenantServiceRegistry);
                 } catch (Exception e) {
                     log.info("[{}] Failed to initialize the service ", label, e);
                     try {
-                        service.destroy();
+                        if (service != null) {
+                            service.destroy();
+                        }
                     } catch (Exception exc) {
                         log.info("[{}] Failed to stop the service ", label, exc);
                     }
@@ -73,23 +77,26 @@ public class DefaultTenantManagerService implements TenantManagerService {
             } else {
                 String label = configuration.getLabel();
                 log.info("[{}] Initializing gateway", configuration.getLabel());
-                GatewayService service = new MqttGatewayService(configuration, c -> {});
+                GatewayService service = null;
                 try {
-                    service.init();
-                    ExtensionServiceCreation serviceCreation = new TenantServicesRegistry(service);
+                    TenantServiceRegistry tenantServiceRegistry = new TenantServiceRegistry();
+                    service = getGatewayService(configuration, c -> {});
+                    tenantServiceRegistry.setService(service);
                     for (TbExtensionConfiguration extensionConfiguration : configuration.getExtensions()) {
                         log.info("[{}] Initializing extension: [{}]", configuration.getLabel(), extensionConfiguration.getType());
-                        ExtensionService extension = serviceCreation.createExtensionServiceByType(service, extensionConfiguration.getType());
+                        ExtensionService extension = tenantServiceRegistry.createExtensionServiceByType(service, extensionConfiguration.getType());
                         extension.init(extensionConfiguration, isRemoteConfiguration);
                         if (extensionConfiguration.getType().equals("HTTP")) {
                             httpServices.add((HttpService) extension);
                         }
                     }
-                    gateways.put(label, (TenantServicesRegistry) serviceCreation);
+                    gateways.put(label, (TenantServiceRegistry) tenantServiceRegistry);
                 } catch (Exception e) {
                     log.info("[{}] Failed to initialize the service ", label, e);
                     try {
-                        service.destroy();
+                        if (service != null) {
+                            service.destroy();
+                        }
                     } catch (Exception exc) {
                         log.info("[{}] Failed to stop the service ", label, exc);
                     }
@@ -98,16 +105,10 @@ public class DefaultTenantManagerService implements TenantManagerService {
         }
     }
 
-    private void onExtensionConfigurationUpdate(String label, String configuration) {
-        TenantServicesRegistry registry = gateways.get(label);
-        log.info("[{}] Updating extension configuration", label);
-        registry.updateExtensionConfiguration(configuration);
-    }
-
     @Override
     public void processRequest(String converterId, String token, String body) throws Exception {
         if (isRemoteConfiguration) {
-            for (TenantServicesRegistry tenant : gateways.values()) {
+            for (TenantServiceRegistry tenant : gateways.values()) {
                 tenant.processRequest(converterId, token, body);
             }
         } else {
@@ -121,7 +122,7 @@ public class DefaultTenantManagerService implements TenantManagerService {
     public void stop() {
         for (String label : gateways.keySet()) {
             try {
-                TenantServicesRegistry registry = gateways.get(label);
+                TenantServiceRegistry registry = gateways.get(label);
                 for (ExtensionService extension : registry.getExtensions().values()) {
                     try {
                         if (isRemoteConfiguration) {
